Moved types into random variable module.
